add KnowledgeSource Arrangement AttendByDays ArrangedDay

+++ src/com/company/Education/Knowledge.java
@@ -5,7 +5,7 @@ public abstract class Knowledge {
     double knowledgeLevel;
     double experienceLevel;
 
-    public Knowledge(int knowledge, int experience) {
+    public Knowledge(double knowledge, double experience) {
         this.knowledgeLevel = knowledge;
         this.experienceLevel = experience;
     }
